Changed types in wrapper to be platform independent (Status is always 32 bit)

+++ pyvisa/ctwrapper/types.py
@@ -37,9 +37,9 @@ def _type_triplet(ctypes_type):
     return _type_pair(ctypes_type) + (_ctypes.POINTER(ctypes_type),)
 
 ViUInt64, ViPUInt64, ViAUInt64    = _type_triplet(_ctypes.c_uint64)
-ViInt64, ViPInt64, ViAInt64       = _type_triplet(_ctypes.c_long)
-ViUInt32, ViPUInt32, ViAUInt32    = _type_triplet(_ctypes.c_ulong)
-ViInt32, ViPInt32, ViAInt32       = _type_triplet(_ctypes.c_long)
+ViInt64, ViPInt64, ViAInt64       = _type_triplet(_ctypes.c_int64)
+ViUInt32, ViPUInt32, ViAUInt32    = _type_triplet(_ctypes.c_uint32)
+ViInt32, ViPInt32, ViAInt32       = _type_triplet(_ctypes.c_int32)
 ViUInt16, ViPUInt16, ViAUInt16    = _type_triplet(_ctypes.c_ushort)
 ViInt16, ViPInt16, ViAInt16       = _type_triplet(_ctypes.c_short)
 ViUInt8, ViPUInt8, ViAUInt8       = _type_triplet(_ctypes.c_ubyte)
